+++ src/Application/Command/Participant/Import.php
@@ -10,23 +10,21 @@
 
 namespace Proximum\Vimeet\Application\Command\Participant;
 
+use Proximum\Vimeet\Application\Command\Command;
 use Proximum\Vimeet\Domain\Model\Type;
 use Symfony\Component\HttpFoundation\File\UploadedFile;
 
-class Import
+class Import implements Command
 {
-    /**
-     * @var Type
-     */
+    /** @var Type */
     public $type;
 
-    /**
-     * @var UploadedFile
-     */
+    /** @var UploadedFile */
     public $file;
 
-    /**
-     * @var string
-     */
+    /** @var string */
     public $charset;
+
+    /** @var bool */
+    public $allowMultiSheet;
 }