+++ src/Infrastructure/Bundle/InfrastructureBundle/EventListener/Security/AuthenticationFailureSubscriber.php
@@ -10,8 +10,8 @@ use Proximum\Vimeet\Domain\Repository\EventRepositoryInterface;
 use Proximum\Vimeet\Domain\Repository\UserRepositoryInterface;
 use Symfony\Component\EventDispatcher\EventSubscriberInterface;
 use Symfony\Component\HttpFoundation\RequestStack;
-use Symfony\Component\Security\Core\Authentication\Token\UsernamePasswordToken;
 use Symfony\Component\Security\Core\AuthenticationEvents;
+use Symfony\Component\Security\Core\Authentication\Token\UsernamePasswordToken;
 use Symfony\Component\Security\Core\Event\AuthenticationFailureEvent;
 
 class AuthenticationFailureSubscriber implements EventSubscriberInterface
@@ -56,7 +56,7 @@ class AuthenticationFailureSubscriber implements EventSubscriberInterface
     {
         $token = $authenticationFailureEvent->getAuthenticationToken();
 
-        if (!$token instanceof UsernamePasswordToken || 'main' !== $token->getProviderKey()) {
+        if (!$token instanceof UsernamePasswordToken || $token->getProviderKey() !== 'main') {
             return;
         }