+++ src/Ui/Bundle/AdminBundle/Controller/Happening/ExportParticipantAction.php
@@ -1,24 +1,14 @@
 <?php
 
-/*
- * This file is part of the Proximum Vimeet project.
- *
- * Copyright (C) Proximum
- *
- * @author Elao <contact@elao.com>
- */
-
 namespace Proximum\Vimeet\Ui\Bundle\AdminBundle\Controller\Happening;
 
 use Proximum\Vimeet\Application\Adapter\AuthorizationCheckerAdapterInterface;
-use Proximum\Vimeet\Application\Adapter\QueryBusInterface;
+use Proximum\Vimeet\Application\Adapter\CommandBusInterface;
 use Proximum\Vimeet\Application\Adapter\RouterInterface;
-use Proximum\Vimeet\Application\Adapter\SerializerAdapterInterface;
+use Proximum\Vimeet\Application\Command\Happening\Export\ScheduleExport;
 use Proximum\Vimeet\Application\Exception\Happening\EmptyHappeningParticipationException;
-use Proximum\Vimeet\Application\Query\Happening\Admin\HappeningParticipantExportViewQuery;
-use Proximum\Vimeet\Application\Serializer\Charset;
 use Proximum\Vimeet\Domain\Model\Event;
-use Proximum\Vimeet\Infrastructure\Bundle\InfrastructureBundle\HttpFoundation\Response\CsvFileResponse;
+use Proximum\Vimeet\Ui\Bundle\AdminBundle\ValueResolver\AdminDomain;
 use Symfony\Component\HttpFoundation\RedirectResponse;
 use Symfony\Component\HttpFoundation\Request;
 use Symfony\Component\HttpFoundation\Session\Flash\FlashBagInterface;
@@ -35,68 +25,39 @@ class ExportParticipantAction
     /** @var RouterInterface */
     private $router;
 
-    /** @var SerializerAdapterInterface */
-    private $serializer;
+    /** @var CommandBusInterface */
+    private $commandBus;
 
-    /** @var QueryBusInterface */
-    private $queryBus;
-
-    /**
-     * @param AuthorizationCheckerAdapterInterface $authorizationCheckerAdapter
-     * @param FlashBagInterface                    $flashBag
-     * @param RouterInterface                      $router
-     * @param QueryBusInterface                    $queryBus
-     * @param SerializerAdapterInterface           $serializerAdapter
-     */
     public function __construct(
         AuthorizationCheckerAdapterInterface $authorizationCheckerAdapter,
         FlashBagInterface $flashBag,
         RouterInterface $router,
-        QueryBusInterface $queryBus,
-        SerializerAdapterInterface $serializerAdapter
+        CommandBusInterface $commandBus
     ) {
         $this->authorizationCheckerAdapter = $authorizationCheckerAdapter;
         $this->flashBag = $flashBag;
-        $this->queryBus = $queryBus;
         $this->router = $router;
-        $this->serializer = $serializerAdapter;
+        $this->commandBus = $commandBus;
     }
 
-    /**
-     * @param Request $request
-     * @param Event   $event
-     *
-     * @throws AccessDeniedException
-     *
-     * @return RedirectResponse|CsvFileResponse
-     */
-    public function __invoke(Request $request, Event $event)
+    public function __invoke(Request $request, Event $event, AdminDomain $adminDomain): RedirectResponse
     {
         if (!$this->authorizationCheckerAdapter->isGranted('PERMISSION_EVENT_ACCESS', $event)) {
             throw new AccessDeniedException('Access Denied!');
         }
 
         try {
-            $happeningParticipantViews = $this->queryBus->handle(
-                new HappeningParticipantExportViewQuery($event, $event->getAvailableLocale($request->getLocale()))
+            $this->commandBus->handle(
+                new ScheduleExport($event, $adminDomain->getAdmin(), $event->getAvailableLocale($request->getLocale()))
             );
+
+            $this->flashBag->add('success', 'flash.admin.happening.participation.export_scheduled');
         } catch (EmptyHappeningParticipationException $exception) {
             $this->flashBag->add('error', 'flash.admin.happening.participation.empty');
-
-            return new RedirectResponse(
-                $this->router->generate('admin_happening_list', ['event' => $event->getId()])
-            );
         }
 
-        $exportedContent = $this->serializer->serialize($happeningParticipantViews, 'csv', [
-            'locale'        => $event->getAvailableLocale($request->getLocale()),
-            'charset'       => Charset::WINDOWS_1252,
-            'csv_delimiter' => ';',
-        ]);
-
-        return new CsvFileResponse(
-            $exportedContent,
-            'export_happening_participants_' . date('Y_m_d_His') . '.csv'
+        return new RedirectResponse(
+            $this->router->generate('admin_happening_list', ['event' => $event->getId()])
         );
     }
 }