+++ tests/Domain/Sheet/Catalog/CanSeeOtherSheetsTest.php
@@ -6,7 +6,7 @@ use PHPUnit\Framework\TestCase;
 use Proximum\Vimeet\Domain\Model\Event;
 use Proximum\Vimeet\Domain\Model\Rule;
 use Proximum\Vimeet\Domain\Model\Sheet;
-use Proximum\Vimeet\Domain\Model\WhoInterface;
+use Proximum\Vimeet\Domain\Model\Type;
 use Proximum\Vimeet\Domain\Repository\RuleRepositoryInterface;
 use Proximum\Vimeet\Domain\Sheet\Catalog\CanSeeOtherSheets;
 
@@ -15,9 +15,9 @@ class CanSeeOtherSheetsTest extends TestCase
     public function testIsSatisfiedBy()
     {
         $event = $this->prophesize(Event::class);
-        $who = $this->prophesize(WhoInterface::class);
+        $who = $this->prophesize(Type::class);
         $rule = $this->prophesize(Rule::class);
-        
+
         $sheet = $this->prophesize(Sheet::class);
         $sheet->getEvent()
             ->shouldBeCalled()
@@ -27,21 +27,21 @@ class CanSeeOtherSheetsTest extends TestCase
             ->shouldBeCalled()
             ->willReturn($who->reveal());
         ;
-        
+
         $ruleRepository = $this->prophesize(RuleRepositoryInterface::class);
         $ruleRepository->getByEventAndSeer($event->reveal(), $who->reveal())
             ->shouldBeCalled()
             ->willReturn($rule->reveal());
-        
+
         $hasAccessToCatalog = new CanSeeOtherSheets($ruleRepository->reveal());
         $this->assertTrue($hasAccessToCatalog->isSatisfiedBy($sheet->reveal()));
     }
-    
+
     public function testIsNotSatisfiedBy()
     {
         $event = $this->prophesize(Event::class);
-        $who = $this->prophesize(WhoInterface::class);
-        
+        $who = $this->prophesize(Type::class);
+
         $sheet = $this->prophesize(Sheet::class);
         $sheet->getEvent()
             ->shouldBeCalled()
@@ -51,7 +51,7 @@ class CanSeeOtherSheetsTest extends TestCase
             ->shouldBeCalled()
             ->willReturn($who->reveal());
         ;
-        
+
         $ruleRepository = $this->prophesize(RuleRepositoryInterface::class);
         $ruleRepository->getByEventAndSeer($event->reveal(), $who->reveal())
             ->shouldBeCalled()